# Use String.build instead of IO::Memory.