--- a/README.md
+++ b/README.md
@@ -1,9 +1,12 @@
 # base32
-
-[![Build Status](https://travis-ci.org/noumar/base32.svg?branch=master)](https://travis-ci.org/noumar/base32)
-[![Docs](http://docrystal.org/badge.svg?style=round)](http://docrystal.org/github.com/noumar/base32)
+[![Crystal CI](https://github.com/didactic-drunk/base32/actions/workflows/crystal.yml/badge.svg)](https://github.com/didactic-drunk/base32/actions/workflows/crystal.yml)
+[![GitHub release](https://img.shields.io/github/release/didactic-drunk/base32.svg)](https://github.com/didactic-drunk/base32/releases)
+![GitHub commits since latest release (by date) for a branch](https://img.shields.io/github/commits-since/didactic-drunk/base32/latest)
+[![Docs](https://img.shields.io/badge/docs-available-brightgreen.svg)](https://didactic-drunk.github.io/base32/master)
 
 Provides encoding and decoding of base32 and base32hex as defined in RFC 4648.
+
+Maintained here temporarily (or permanently) until @noumar returns.
 
 ## Installation
 
@@ -12,7 +15,7 @@ Add this to your application's `shard.yml`:
 ```yaml
 dependencies:
   base32:
-    github: noumar/base32
+    github: didactic-drunk/base32
 ```
 
 ## Usage
@@ -33,7 +36,7 @@ TODO: Write development instructions here
 
 ## Contributing
 
-1. Fork it ( https://github.com/noumar/base32/fork )
+1. Fork it ( https://github.com/didactic-drunk/base32/fork )
 2. Create your feature branch (git checkout -b my-new-feature)
 3. Commit your changes (git commit -am 'Add some feature')
 4. Push to the branch (git push origin my-new-feature)
@@ -41,4 +44,5 @@ TODO: Write development instructions here
 
 ## Contributors
 
-- [noumar](https://github.com/noumar) / Mikael Karlsson - creator, maintainer
+- [noumar](https://github.com/noumar) / Mikael Karlsson - creator
+- [didactic-drunk](https://github.com/didactic-drunk) - current maintainer