--- a/spec/base32_spec.cr
+++ b/spec/base32_spec.cr
@@ -12,6 +12,9 @@ describe Base32 do
     Base32.encode("foobar").should eq("MZXW6YTBOI======")
 
     Base32.encode("Hello World!").should eq("JBSWY3DPEBLW64TMMQQQ====")
+
+    # Compared with Ruby version
+    Base32.encode(Bytes.new(4)).should eq("AAAAAAA=")
   end
 
   it "encodes to base32 (w/o padding)" do
@@ -90,9 +93,16 @@ describe Base32 do
     Base32.hex_decode_string("CPNMUOJ1E8").should eq("foobar")
   end
 
-  it "encodes/decodes with NUL bytes" do
+  describe "NUL byte" do
     str = "\0foo\0bar\0"
-    str2 = Base32.decode_string(Base32.encode(str))
-    str2.should eq str
+    bstr = "ABTG63YAMJQXEAA="
+
+    it "encode" do
+      Base32.encode(str).should eq bstr
+    end
+
+    it "decode" do
+      Base32.decode_string(bstr).should eq str
+    end
   end
 end