--- a/src/base32.cr
+++ b/src/base32.cr
@@ -32,65 +32,57 @@ module Base32
 
   # :nodoc:
   private def to_base32(data, pad : Bool, chars : Array(Char)) : String
-    mio = IO::Memory.new((data.bytesize // 5) * 8)
-
-    data.to_slice.each_slice(5) do |slice|
-      bits = 0_u64
-      0.to(slice.size - 1) do |j|
-        bits = bits | (slice[j].to_u64 << (4 - j)*8)
+    ssize = data.bytesize * 8 // 5
+    ssize += 7 if pad
+    String.build(ssize) do |sb|
+      bits = 0
+      b = 0
+      data.to_slice.each do |c|
+        b = (b << 8) | c
+        bits += 8
+        while bits >= 5
+          bits -= 5
+          sb << chars[b >> bits]
+          mask = (1 << bits) - 1
+          b &= mask
+        end
       end
 
-      mask = 0x1F_u64 << 35
-      7.to(0) do |i|
-        num = (bits & mask) >> i*5
-        mio << chars[num]
-        mask = (mask >> 5)
+      if bits > 0
+        sb << chars[b << (5 - bits)]
+      end
+
+      if pad
+        rem = sb.bytesize % 8
+        (8 - rem).times { sb << PAD } if rem > 0
       end
     end
-
-    # Exclude empty trailing chars
-    while mio.buffer[mio.pos - 1] == chars[0].ord
-      mio.pos -= 1
-    end
-
-    # Fill with padding
-    until (mio.pos % 8 == 0)
-      mio << PAD
-    end
-
-    # Exclude padding if not wanted
-    sl = mio.to_slice
-    while sl[-1] == PAD.ord
-      sl = sl[0, sl.size - 1]
-    end if sl.size > 0 && pad == false
-
-    String.new(sl)
   end
 
+  IGNORE_CHARS = ['\n'.ord, '\r'.ord, '='.ord]
+
   # :nodoc:
-  private def from_base32(data, map : Hash(Char, Int)) : Slice(UInt8)
+  private def from_base32(data, map : Hash(Char, Int)) : Bytes
     mio = IO::Memory.new((data.bytesize // 8) * 5)
 
-    data.to_slice.select { |s| !['\n'.ord, '\r'.ord, '='.ord].includes?(s) }.each_slice(8) do |slice|
+    data.to_slice.select { |s| !IGNORE_CHARS.includes?(s) }.each_slice(8) do |slice|
       bits = 0_u64
-      0.to(slice.size - 1) do |j|
-        bits = bits | (map[slice[j].chr].to_u64 << (7 - j)*5)
+      slice.each_with_index do |b, j|
+        bits = bits | (map[b.chr].to_u64 << (7 - j)*5)
       end
 
       mask = 0xFF_u64 << 32
-      4.to(0) do |i|
+
+      rem_bytes = slice.size * 5 // 8
+
+      4.downto(5 - rem_bytes) do |i|
         num = (bits & mask) >> i*8
         mio.write_byte(num.to_u8)
         mask = (mask >> 8)
       end
     end
 
-    # Exclude trailing zero bytes
-    sl = mio.to_slice
-    while sl[-1] == 0
-      sl = sl[0, sl.size - 1]
-    end if sl.size > 0
-    sl
+    mio.to_slice
   end
 
   # Encode data as base32 with padding, or without if `pad` = false
@@ -99,7 +91,7 @@ module Base32
   end
 
   # Decode base32 data, regardless if padded or not
-  def decode(data) : Slice(UInt8)
+  def decode(data) : Bytes
     from_base32(data, DEC_STD)
   end
 
@@ -114,7 +106,7 @@ module Base32
   end
 
   # Decode base32hex data, regardless if padded or not
-  def hex_decode(data) : Slice(UInt8)
+  def hex_decode(data) : Bytes
     from_base32(data, DEC_HEX)
   end