New faster implementation of Base32.encode.
Correctly handle NUL bytes.
Add NUL spec.